requirements.txt
Django
psycopg2
djangorestframework
django_filters
djangorestframework-filters
httpie